fix(router): use tsx for unified routes
