+++ src/router/index.tsx
@@ -1,6 +1,6 @@
 
 import { createBrowserRouter } from 'react-router-dom';
-import { buildUnifiedRoutes } from './buildUnifiedRoutes';
+import { buildUnifiedRoutes } from './buildUnifiedRoutes.tsx';
 
 export const router = createBrowserRouter(buildUnifiedRoutes(), {
   basename: import.meta.env.BASE_URL ?? '/',
